docker/requirements.txt
pyaudio
vosk
mysql-connector-python
requests
python-dotenv
flask